# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des utilisateurs et vendeurs.",
        "title": "o'z api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support API",
            "email": "goliyaohugues@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authentifie l'utilisateur et retourne un token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion utilisateur",
                "parameters": [
                    {
                        "description": "Identifiant et mot de passe",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ex: {user: {}, token: '...'}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Identifiants incorrects",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalide la session (côté client). Retourne un message de confirmation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "responses": {
                    "200": {
                        "description": "message: Déconnexion réussie",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Crée un nouveau compte utilisateur et retourne un token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription d'un utilisateur",
                "parameters": [
                    {
                        "description": "Données d'inscription",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ex: {message: '...', user: {}, token: '...'}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Payload ou validation invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permet à l'utilisateur connecté de modifier son mot de passe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Changement de mot de passe",
                "parameters": [
                    {
                        "description": "Ancien et nouveau mot de passe",
                        "name": "passwords",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.resetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Mot de passe mis à jour",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Ancien mot de passe incorrect",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "transport.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "password",
                "phone",
                "role",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jean@zoro.com"
                },
                "firstname": {
                    "type": "string",
                    "example": "Jean"
                },
                "lastname": {
                    "type": "string",
                    "example": "Dupont"
                },
                "parentId": {
                    "type": "integer"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "pass1234"
                },
                "phone": {
                    "type": "string",
                    "example": "0708091011"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "vendeur"
                    ],
                    "example": "vendeur"
                },
                "username": {
                    "type": "string",
                    "minLength": 3,
                    "example": "jean_vendeur"
                }
            }
        },
...
        "transport.loginRequest": {
            "type": "object",
            "required": [
                "identifier",
                "password"
            ],
            "properties": {
                "identifier": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "transport.resetPasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 6
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" suivi d'un espace et du token JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}